#6 Fix TSLint errors: Add javadoc comments to methods
