+++ src/index.ts
@@ -1,2 +1,2 @@
-export { ReportOptions } from './models/reportOptions';
+export { IReportOptions as ReportOptions } from './models/reportOptions';
 export { Reporter } from './reporter';
