+++ src/models/cucumberReportSummary.ts
@@ -1,7 +1,16 @@
-import { IResult, ResultStatus } from './result';
+import { IResult } from './result';
+import { ResultStatus } from './resultStatus';
 
-// Can be aggregated at the Feature level, and then aggregated at the report level
-
+/**
+ * The Cucumber Report Summary maintains a summary of all child
+ * elements in a cucumber report.
+ * - For a suite of features, this summary is the number of passing/failing
+ * features that belong to the suite.
+ * - For a single Feature, this summary is the number of passing/failing
+ * scenarios that belong to the feature.
+ * - For a single Scenario, this summary is the number of passing/failing
+ * steps that belong to the scenario
+ */
 export class CucumberReportSummary {
   public passed: number = 0;
   public failed: number = 0;
@@ -28,6 +37,12 @@ export class CucumberReportSummary {
   get isFailed(): boolean { return this.failed > 0; }
   get isPassed(): boolean { return this.passed === this.total; }
 
+  /**
+   * Updates this summary to include the results of the child summary.
+   * The child summary is rolled up into this parent summary
+   * @param child The Cucumber Report summary for the sub-element
+   *  (Feature in a Suite, Scenario in a single Feature, Step in a single Scenario)
+   */
   public aggregateChildSummary(child: CucumberReportSummary): void {
     this.totalDuration += child.totalDuration;
 
@@ -40,6 +55,12 @@ export class CucumberReportSummary {
     this.unknown += child.unknown;
   }
 
+  /**
+   * Updates the summary based on the raw cucumber report step status.
+   * Performs the base level summarisation of the report at the lowest leaf
+   * of the report tree (the Step)
+   * @param result The Step result from the raw Cucumber report
+   */
   public updateFromReportResult(result: IResult): void {
     this.totalDuration += <number> (result.duration | 0);
