+++ src/models/featureSummary.ts
@@ -1,5 +1,10 @@
 import { CucumberReportSummary } from './cucumberReportSummary';
 
+/**
+ * Aggregated summary of a single Feature.
+ * Contains summary of each Scenario that belongs to this
+ * feature
+ */
 export interface IFeatureSummary {
   scenarios: CucumberReportSummary[];
   featureSummary: CucumberReportSummary;
