+++ src/models/reportOptions.ts
@@ -1,9 +1,13 @@
 
-export interface ReportOptions {
+/**
+ * Defines the options that can be provided to the YACHR
+ * Cucumber report generator
+ */
+export interface IReportOptions {
   /** HTML file output path
    * @Example
    * `./e2e-reports/report.html`
-  */
+   */
   output: string;
 
   /** Path to Cucumber file in JSON format */
