+++ src/models/resultStatus.ts
@@ -1,0 +1,11 @@
+/**
+ * Lists the possible result statuses for a step result in the Cucumber
+ * report
+ */
+export enum ResultStatus {
+  passed = 'passed',
+  failed = 'failed',
+  undefined = 'undefined',
+  pending = 'pending',
+  ambiguous = 'ambiguous'
+}
