+++ src/models/tag.ts
@@ -1,3 +1,6 @@
+/**
+ * Defines a tag as defined in the Gherkin specification
+ */
 export interface ITag {
   name: string;
   line: number;
