+++ src/reportAggregator.ts
@@ -20,6 +20,8 @@ export class ReportAggregator {
    * @param suite Array of cucumber results. Standard output from a cucumber test being run.
    */
   public getSummaryForSuite(suite: ICucumberResult[]): ISuiteSummary {
+    // TODO: I've defined a datamodel for a FeatureSuite. Update code to reflect this
+
     const response: ISuiteSummary = {
       features: [],
       suiteSummary: new CucumberReportSummary()
@@ -39,6 +41,10 @@ export class ReportAggregator {
     return response;
   }
 
+  /**
+   * Aggregates a single feature. Aggregates all scenarios in the feature to generate a summary.
+   * @param feature The feature to aggregate
+   */
   public getSummaryForFeature(feature: ICucumberResult): IFeatureSummary {
     const response: IFeatureSummary = { scenarios: [], featureSummary: new CucumberReportSummary(), featureName: feature.name };
 
@@ -51,6 +57,10 @@ export class ReportAggregator {
     return response;
   }
 
+  /**
+   * Aggregates a single scenario. Aggregates the results of all steps in the Scenario to generate a summary.
+   * @param scenario The scenario to aggregate
+   */
   public getSummaryForScenario(scenario: IElement): CucumberReportSummary {
     const scenarioSummary = new CucumberReportSummary();
 
@@ -64,6 +74,10 @@ export class ReportAggregator {
     return scenarioSummary;
   }
 
+  /**
+   * Summarises the result of a single Gherkin Step
+   * @param step The Cucumber Test result of the Gherkin step
+   */
   private summariseStep(step: IStep): CucumberReportSummary {
     const stepSummary = new CucumberReportSummary();
