+++ src/reporter.ts
@@ -3,19 +3,30 @@ import * as Handlebars from 'handlebars';
 
 import { CucumberReportSummary } from './models/cucumberReportSummary';
 import { ICucumberResult } from './models/cucumberResult';
-import { HtmlModel } from './models/htmlModel';
-import { ReportOptions } from './models/reportOptions';
+import { IHtmlModel } from './models/htmlModel';
+import { IReportOptions } from './models/reportOptions';
 import { ReportAggregator } from './reportAggregator';
 
+/**
+ * The main YACHR Cucumber HTML Report generator.
+ * Provides an API to read in a Cucumber Test Report,
+ * and generate a static HTML page which summarises the report
+ * into a dashboard
+ */
 export class Reporter {
-  public generate(options: ReportOptions): void {
+
+  /**
+   * Generates the HTML Report summary for the Cucumber Test Report.
+   * @param options Options to configure how the reporter will generate the HTML report
+   */
+  public generate(options: IReportOptions): void {
     options = this.populateDefaultOptionsIfMissing(options);
 
     const results = this.parseJsonFile(options.jsonFile);
 
     const aggregator = new ReportAggregator();
 
-    const data = <HtmlModel> {
+    const data = <IHtmlModel> {
       cucumberReportSummary: aggregator.getSummaryForSuite(results),
       cucumberResult: results
     };
@@ -48,23 +59,30 @@ export class Reporter {
     }
   }
 
-  /** Used by generate to add in any default options that need to overwrite empty parameters */
-  private populateDefaultOptionsIfMissing(options: ReportOptions): ReportOptions {
-    const defaultOptions = <ReportOptions> {
-      htmlTemplate: __dirname + '/templates/standard.html'
-    };
-
-    return {...defaultOptions, ... options};
-  }
-
+  /**
+   * Parses a JSON file (at the given file path) into the data model for
+   * the Cucumber Report. This report can then be used with strong typing
+   * when accessing components of the report
+   * @param resultsFile The path to the Cucumber Test Results file
+   */
   public parseJsonFile(resultsFile: string): ICucumberResult[] {
-    let results: ICucumberResult[];
-    let input: string;
     try {
       return JSON.parse(fs.readFileSync(resultsFile, 'utf8'));
     } catch (err) {
       console.error('Error reading file: ' + resultsFile);
       throw (err);
     }
+  }
+
+  /**
+   * Used by generate to add in any default options that need to overwrite empty parameters
+   * @param options The options as passed in by the user
+   */
+  private populateDefaultOptionsIfMissing(options: IReportOptions): IReportOptions {
+    const defaultOptions = <IReportOptions> {
+      htmlTemplate: __dirname + '/templates/standard.html'
+    };
+
+    return {...defaultOptions, ... options};
   }
 }
